+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- [TD-6981] Add logic to indexed Structures updating dates
+  - updated_at takes Data Structure date
+  - new field last_change_at takes Structure or Version most recient date 
+
 ## [6.16.0] 2024-12-16
 
 ### Added